+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Optionally add SH/SHUI managed classes to shapes graph
 - Optionally import CMEM queryu cayalog to shapes graph
+- Optionally add target class depiction to node shape if present in the data graph or the ontology.
+For the ontology graph name the target class namespace is considered, with or without trailing separator.
 
 
 ## [3.0.1] 2025-03-13